pyproject.toml
[tool.poetry]
name = "llme"
version = "0.1.0"
description = ""
authors = ["David Minarsch <david.minarsch@googlemail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "0.0.141"
openai = "0.27.2"
google-search-results = "^2.4.1"
flask = "^2.2.3"
python-dotenv = "^0.14.0"
llama-index = "^0.4.39"
pinecone-client = "^2.2.1"
ruff = "^0.0.260"
black = "*"
open-aea = {version = "1.33.0", extras = ["all"]}
open-aea-ledger-ethereum = {version = "1.32.0"}
unstructured = {extras = ["local-inference"], version = "^0.5.12"}
tiktoken = "^0.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"